Dried Hide highlight

+++ src/haven/WItem.java
@@ -222,12 +222,15 @@ public class WItem extends Widget implements DTarget {
 		item.wdgmsg("take", c);
 	    return(true);
 	} else if(btn == 3) {
-		if (ui.modshift) {
-			wdgmsg("transfer_all", item.resource().name);
-		} else if (ui.modctrl) {
-			wdgmsg("drop_all", item.resource().name);
+		if (Shovel.getSettings().enableGroupHotkeys) {
+			if (ui.modshift) {
+				wdgmsg("transfer_all", item.resource().name);
+			} else if (ui.modctrl) {
+				wdgmsg("drop_all", item.resource().name);
+			} else
+				item.wdgmsg("iact", c, ui.modflags());
 		} else
-	    	item.wdgmsg("iact", c, ui.modflags());
+			item.wdgmsg("iact", c, ui.modflags());
 	    return(true);
 	}
 	return(false);
